resolved pb with no planes when sprite beside a wall in the map parser

+++ libft/srcs/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: lwourms <lwourms@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/23 14:19:15 by lwourms           #+#    #+#             */
-/*   Updated: 2021/05/16 20:37:46 by lwourms          ###   ########.fr       */
+/*   Updated: 2021/05/26 17:36:48 by lwourms          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
